int -> Py_ssize_t in cpython pxd files

+++ Cython/Includes/cpython/tuple.pxd
@@ -24,10 +24,10 @@ cdef extern from "Python.h":
     # pointing to Python objects. "PyTuple_Pack(2, a, b)" is
     # equivalent to "Py_BuildValue("(OO)", a, b)".
 
-    int PyTuple_Size(object  p) except -1
+    Py_ssize_t PyTuple_Size(object  p) except -1
     # Take a pointer to a tuple object, and return the size of that tuple.
 
-    int PyTuple_GET_SIZE(object  p)
+    Py_ssize_t PyTuple_GET_SIZE(object  p)
     # Return the size of the tuple p, which must be non-NULL and point
     # to a tuple; no error checking is performed.
